+++ RELEASES.md
@@ -2,6 +2,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.2.5-SNAPSHOT`
 * prov-template-archetype: navigator tab supports backward and forward traversals
+* service-templates-core: idempotency of statement posting, and support for idempotency-key in template service
 
 `2.2.4`